DOC #279 synchronize API & docs with prior version
