+++ hkl/tests/test_configuration.py
@@ -4,7 +4,8 @@ import pytest
 
 from .. import DiffractometerConfiguration
 from ..configuration import EXPORT_FORMATS
-from ..configuration import REQUIRED_CONFIGURATION_KEYS_TYPES
+
+# from ..configuration import REQUIRED_CONFIGURATION_KEYS_TYPES
 
 
 def test_e4cv(e4cv):
@@ -59,32 +60,32 @@ def test_format(fmt, e4cv):
         config.restore(cfg)  # test restore with automatic type recognition
 
 
-@pytest.mark.parametrize("action", "rm set".split())  # remove or set keys incorrectly
-@pytest.mark.parametrize(
-    "key, value, failure",
-    [[k, object, AssertionError] for k in REQUIRED_CONFIGURATION_KEYS_TYPES]
-)
-def test_validation_fails(action, key, value, failure, tardis):
-    assert len(tardis.calc._samples) == 1
-    assert tardis.calc.sample.name == "main"
+# @pytest.mark.parametrize("action", "rm set".split())  # remove or set keys incorrectly
+# @pytest.mark.parametrize(
+#     "key, value, failure",
+#     [[k, object, AssertionError] for k in REQUIRED_CONFIGURATION_KEYS_TYPES]
+# )
+# def test_validation_fails(action, key, value, failure, tardis):
+#     assert len(tardis.calc._samples) == 1
+#     assert tardis.calc.sample.name == "main"
 
-    with pytest.raises(TypeError):
-        cfg = DiffractometerConfiguration("wrong diffractometer object")
-        cfg.validate_config_dict({})
+#     with pytest.raises(TypeError):
+#         cfg = DiffractometerConfiguration("wrong diffractometer object")
+#         cfg.validate_config_dict({})
 
-    with pytest.raises(TypeError):
-        cfg = DiffractometerConfiguration(tardis)
-        cfg.validate_config_dict("wrong configuration object")
+#     with pytest.raises(TypeError):
+#         cfg = DiffractometerConfiguration(tardis)
+#         cfg.validate_config_dict("wrong configuration object")
 
-    with pytest.raises(failure):
-        cfg = DiffractometerConfiguration(tardis)
-        assert isinstance(cfg, dict), f"{cfg}"
+#     with pytest.raises(failure):
+#         cfg = DiffractometerConfiguration(tardis)
+#         assert isinstance(cfg, dict), f"{cfg}"
 
-        if action == "rm":
-            cfg.pop(key)
-        elif action == "set":
-            cfg[key] = value
-        cfg.validate_config_dict(cfg)
+#         if action == "rm":
+#             cfg.pop(key)
+#         elif action == "set":
+#             cfg[key] = value
+#         cfg.validate_config_dict(cfg)
 
 
 # TODO: test sample dictionary
